+++ examples/kyb-app/src/components/organisms/DynamicUI/hooks/useRuleExecutor/useRuleExecutor.ts
@@ -9,7 +9,7 @@ import { Rule, UIElement } from '@app/domains/collection-flow';
 import { AnyObject } from '@ballerine/ui';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import { JmespathRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/jmespath.rule-engine';
-import { DestinationRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/destination.rule-engine';
+import { DocumentsRuleEngine } from '@app/components/organisms/DynamicUI/rule-engines/documents.rule-engine';
 
 export const useRuleExecutor = (
   context: AnyObject,
@@ -31,7 +31,7 @@ export const useRuleExecutor = (
         new JsonLogicRuleEngine(),
         new JsonSchemaRuleEngine(),
         new JmespathRuleEngine(),
-        new DestinationRuleEngine(),
+        new DocumentsRuleEngine(),
       ]),
     [],
   );
@@ -60,6 +60,12 @@ export const useRuleExecutor = (
   useEffect(() => {
     executeRules(context, rules, definition, uiStateRef.current);
   }, [context, rules, uiStateRef, definition, executeRules]);
+
+  if (import.meta.env.MODE === 'development') {
+    if (executionResult.length && executionResult.every(r => !r.isValid && r.errors.length)) {
+      console.log('Rules execution result', executionResult);
+    }
+  }
 
   return executionResult;
 };