+++ examples/kyb-app/src/components/organisms/DynamicUI/types.ts
@@ -1,6 +1,7 @@
 import {
   UIElementStateSetter,
   UILoadingSetter,
+  UIStateSetter,
 } from '@app/components/organisms/DynamicUI/hooks/useUIStateLogic/hooks/useUIElementsStateLogic/types';
 import { UIState } from '@app/components/organisms/DynamicUI/hooks/useUIStateLogic/types';
 
@@ -9,5 +10,6 @@ export interface DynamicUIContext {
   helpers: {
     setUIElementState: UIElementStateSetter;
     setLoading: UILoadingSetter;
+    overrideState: UIStateSetter;
   };
 }