+++ examples/kyb-app/src/components/organisms/UIRenderer/elements/JSONForm/components/NationalityPicker/NationalityPicker.tsx
@@ -1,15 +1,13 @@
 import { RJSVInputProps, TextInputAdapter } from '@ballerine/ui';
-import nationality from 'i18n-nationality';
-import locale from 'i18n-nationality/langs/en.json';
+import locale from './langs/en.json';
 import { useMemo } from 'react';
-nationality.registerLocale(locale);
 
 export const NationalityPicker = (props: RJSVInputProps) => {
   const options = useMemo(
     () =>
-      Object.keys(nationality.getAlpha2Codes()).map(key => ({
-        const: key,
-        title: nationality.getName(key, 'en'),
+      Object.entries(locale.nationalities).map(([alpha2Code, nationality]) => ({
+        const: alpha2Code,
+        title: nationality,
       })),
     [],
   );