+++ examples/kyb-app/src/domains/collection-flow/types/ui-schema.types.ts
@@ -17,6 +17,7 @@ export interface JSONLogicRule extends BaseRule {
 }
 export interface DocumentsValidatorRule extends BaseRule {
   value: {
+    documentId: string;
     destination: string;
     required: boolean | Rule;
     errorMessage: string;
@@ -49,7 +50,7 @@ export type Rule = JSONLogicRule | JMESPathRule | DocumentsValidatorRule;
 
 export type UIElementDestination = string;
 
-export interface UIElement<TElementParams> {
+export interface UIElement<TElementParams = AnyObject> {
   name: string;
   type: UIElementType;
   availableOn?: Rule[];