+++ services/workflows-service/src/ui-definition/utils/collection-flow/2-business-info-page.ts
@@ -262,12 +262,7 @@ export const BusinessInfoPage = {
                 label: 'Corporate type',
                 jsonFormDefinition: {
                   type: 'string',
-                },
-                uiSchema: {
-                  'ui:placeholder': 'Choose',
-                  'ui:field': 'AutocompleteInput',
-                  'ui:label': true,
-                  options: [
+                  oneOf: [
                     { title: 'Sole Proprietorship', const: 'sole_proprietorship' },
                     { title: 'Partnership', const: 'partnership' },
                     { title: 'Corporation', const: 'corporation' },
@@ -321,7 +316,7 @@ export const BusinessInfoPage = {
           elements: [
             {
               name: 'next-page-button',
-              type: 'json-form:button',
+              type: 'submit-button',
               options: {
                 uiDefinition: {
                   classNames: ['align-right', 'padding-top-10'],