+++ services/workflows-service/src/workflow/workflow.service.ts
@@ -1408,6 +1408,7 @@ export class WorkflowService {
       );
 
     let contextToInsert = structuredClone(context);
+    contextToInsert.entity.ballerineEntityId ||= entityId;
 
     const entityConnect = {
       [`${entityType}Id`]: entityId,
@@ -1658,6 +1659,8 @@ export class WorkflowService {
   ): Promise<TEntityId | null> {
     if (entity.ballerineEntityId) {
       return entity.ballerineEntityId as TEntityId;
+    } else if (!entity.id) {
+      return null;
     } else {
       if (entity.type === 'business') {
         const res = await this.businessRepository.findByCorrelationId(